+++ apps/@sparrow-desktop/src/repositories/workspace.repository.ts
@@ -9,7 +9,7 @@ import { EnvironmentRepository } from "./environment.repository";
 import { TabRepository } from "./tab.repository";
 
 export class WorkspaceRepository {
-  constructor() {}
+  constructor() { }
   private collectionRepository = new CollectionRepository();
   private environmentRepository = new EnvironmentRepository();
   private tabRepository = new TabRepository();
@@ -414,46 +414,37 @@ export class WorkspaceRepository {
   };
 
   public searchWorkspaces = async (searchQuery: string): Promise<WorkspaceDocument[]> => {
-  if (!searchQuery.trim()) {
-    // If search query is empty, return recently updated workspaces
+    if (!searchQuery.trim()) {
+      // If search query is empty, return recently updated workspaces
+      return await RxDB.getInstance()
+        .rxdb.workspace.find({
+          sort: [{ updatedAt: 'desc' }],
+          limit: 3
+        })
+        .exec();
+    }
+
+    // If there's a search query, filter workspaces by name or description
+    const searchRegex = new RegExp(searchQuery, 'i');
     return await RxDB.getInstance()
       .rxdb.workspace.find({
-        sort: [{ updatedAt: 'desc' }],
+        selector: {
+          $or: [
+            { name: { $regex: searchRegex } },
+            { description: { $regex: searchRegex } }
+          ]
+        },
         limit: 3
       })
       .exec();
-  }
-
-  // If there's a search query, filter workspaces by name or description
-  const searchRegex = new RegExp(searchQuery, 'i');
-  return await RxDB.getInstance()
-    .rxdb.workspace.find({
-      selector: {
-        $or: [
-          { name: { $regex: searchRegex } },
-          { description: { $regex: searchRegex } }
-        ]
-      },
-      limit: 3
-    })
-    .exec();
-};
-
-public getRecentWorkspaces = async (limit: number = 5): Promise<WorkspaceDocument[]> => {
-  return await RxDB.getInstance()
-    .rxdb.workspace.find({
-      sort: [{ updatedAt: 'desc' }],
-      limit : 3,
-    })
-    .exec();
-};
-
-public fetchWorkspaceDetails  = async (workspaceId: string): Promise<{teamName: string, workspaceName: string}> => {
-  const workspace = await this.readWorkspace(workspaceId);
-  
-  return {
-    teamName: workspace?.team?.teamName || '',
-    workspaceName: workspace?.name || ''
   };
-};
+
+  public getRecentWorkspaces = async (limit: number = 5): Promise<WorkspaceDocument[]> => {
+    return await RxDB.getInstance()
+      .rxdb.workspace.find({
+        sort: [{ updatedAt: 'desc' }],
+        limit: 3,
+      })
+      .exec();
+  };
 }
