+++ packages/@sparrow-common/src/components/global-search/SuggestionTags.svelte
