+++ packages/@sparrow-common/src/components/global-search/index.ts
