+++ packages/@sparrow-common/src/components/global-search/types.ts
