+++ packages/@sparrow-common/src/components/index.ts
@@ -1,4 +1,3 @@
-
 export { default as Header } from "./header/Header.svelte";
 export { default as LoginBanner } from "./banner/login-banner.svelte";
 export { default as LoginPopup } from "./popup/login-popup.svelte";
@@ -9,3 +8,5 @@ export { default as PulseCircle } from "./pulse-circle/PulseCircle.svelte";
 export { default as WelcomePopUpWeb } from "./welcome-pop-up-web/WelcomePopUpWeb.svelte";
 export { default as DefaultHeader } from "./default-header/DefaultHeader.svelte";
 export { default as LaunchDesktop } from "./lanuch-desktop/LanuchDesktop.svelte";
+export { default as GlobalSearch } from "./global-search/GlobalSearch.svelte";
+export { default as SwitchWorkspace } from "./switch-workspace/SwitchWorkspace.svelte";
